application/requirements.txt
fastapi 
uvicorn 
pillow 
python-multipart 
requests
psutil
mlflow
starlette
keras
tensorflow==2.11.0
tensorflow-addons==0.20.0
torch==2.0.1
torchvision==0.15.2
numpy
scikit-image
dropbox
pydantic
SQLAlchemy